[api] prevent the message sending to clients not listening on, thanks to eventMap and listenerTree
done by issuing a getEvents call on a client, to get the clients event map. Then we recursively walk the event map to find if the client is actually listening on this event before we trigger the request.

It may reduce significantly the network overload, but we add a little bit more computation at server emit time.

All tests should be passing.