[major] [refactor] #78 #76 #60 #53 #52
**Removed the concept of `reserved` hook namespaces which won't rebroadcast.**

In previous versions of hook.io, we required that certain namespaces be reserved in order to prevent circular messaging. It seems with recent additions to the API, we can now safetly remove this reserved namespace.

**First pass at fixing multi-level trees**

Now that we can listen for other hook's internal events, we have better control over spawning multi-level trees and getting the correct `ready` events.