Implement registry of events on the server hook
The server hook now holds in his list of clients the list of the
events they are listening to.

The events are set up on the connection of a client, and updated
every time a client listens or stops listening to an event.
The reconnection process is handled as a normal connection, so all
the events will be sent again.